Implemented an Array in python
C/C++ arrays are native an obvious and can be done in less that three
lines but for python, you gotta vibe to the momemnt to create an array,
simply a fixed size list with slots so that the size is constant. With
that, I have implemented a simple usable array. YAH!