Finished off the set data structure with tests.
I have not really implemented the ty.MutableSet completely but I have
added all operations that are needed for a set, that is, intersection,
set difference and set union. This is not in any way the complete list,
as set product is missing and all the other things implemented by the
pyton set that are not part of the ty.MutableSet interface.
I have had a hard time trying to understand all these set operations and
implementations and I have decided to stop here until a highly unlikely
later time. For now, I want to make progress to the next data structure,
Trees and hopefully, the dreadful Fibonacci Heap. A beast of its kind.