Implemented a test BST
Turns out, the BST is completely different from binary search. For one,
BST is a data structure while binary search is an algorithm. Binary
Search Tree (BST) is very performant with logarithmic insertions and
searching but this can become unbearably slow as the tree is more skewed
(unbalanced). I've cooked up a simple setup to see and experiment with
this amazing data structure and I have to say, it is amazing. A more
capable and clean one will be made later but for now, I need to come up
with a way to find the height of the tree and balancing techniques.
Night.