Implemented and read on Bloom Filters
The theory behind blooms is nothing short but amazing and genius. How
data is serialized into bits, yes bits. Although it is probabilistic, it
is very performant. My implementation is very lazy but the approach is
general enough. Instead of multiple hash functions, I use one
configurable hash function that depending on the setting, can produce
different hash values hence can be wrapped multiple times to achieve
multiple bit activates and reduce the number of false-alarms, though
one should also increase the number of bits in the bloom.

The bloom is one of the advanced data structures I've read on and I
believe it deserves a spot in the extras archive along with others to
come like treaps, d-ary heaps and red-black trees.