Solidified the bloom filter implementation
Made the hash function a little more solid and capable. Tried to fix a
few bugs and added a size to the bloom for tracking confidence
calculation and documented it. Bloom filters are amazing and most
interesting part of it is the idea behind it; magnificent. A little more
research will be conducted on the structure; analysis.