Design and learned a new python trick
I know finished on bloom filters but till now I'm still amazed on their
functionality. With that, I spent a little more time researching on how
to improve my implementation and I got a few answers.

One, increase the size of the bit pool to get a higher capacity and
hence lower false positive.
Two, increase the number of bit markers (number of hash functions).

The first one was interesting as I had stumbled on an error in my
earlier implementation that involved big integers in python. Well on
further analysis of this error I discovered that the error occurred when
I tried to print the integer which is simply converting it to a string
but if the integer has too many digits this would result in a very huge
string which python was preventing, did you see that, this means an
integer can be as big as your machine's memory, I literally created an
integer with over 2 million bits (~244 KB) and I could have kept going
the issue was printing it. Python provides a way to convert a bytes
object to an int and an int to bytes which is even cooler, you can store
this big number in a file or something. What I mean is that you can use
the int as a bit pool and is very reliable, unfortunately, I learned
this after creating the poolsyou will find in `pool.py` although I did
enjoy writing it down from the API to the concrete implementations; this
is the skill I've been trying to improve.

I've also settled down on a counting policy where I only increment the
size of the bloom if the value being inserted returns false when checked
for presence in the bloom. Turns out, it can help calculate the false
positive rate I tried to calculate earlier with the `confidence` property.

I will add the formula later and do some more analysis. Night.