Traversals without recursion: iteration.
There are multiple ways of improving an algorithm, making a recursive
algorithm linear is one of them; using loops instead of recursion. I've
converted the tree traversal algorithms from using recursion to looping
and hopefully made them faster, I'm yet to test this but the
implementations are done. Tests for the treap, randomized treap and the
improved traversals are yet to come, Later.