Naive disjoint set implementation
Using equivalence class properties, we were able to implement a disjoint
set using dict and sets. It does work well but lots of room for
improvement are available. Heuristics such as path comression and using
trees are some of the ways for improvement.
Also cleaned up a little, there was a lot of code repetition all over,
but by packaging the whole repo as one mess, we can allow code reuse and
hence easier testing and maintenance.